+++ lib/zedex/impl/syntax_helpers.ex
@@ -53,4 +53,63 @@ defmodule Zedex.Impl.SyntaxHelpers do
 
   def maybe_add_ann_est(est, nil), do: est
   def maybe_add_ann_est(est, ann), do: :erl_syntax.add_ann(ann, est)
+
+  def replace_applications_in_function(
+        function_ast,
+        original_mfa,
+        replacer_fn
+      ) do
+    map_typed(
+      function_ast,
+      fn
+        :application, node ->
+          case application_est_to_mfa(node) do
+            {:ok, ^original_mfa} -> replacer_fn.(node)
+            _ -> node
+          end
+
+        _, node ->
+          node
+      end
+    )
+  end
+
+  def map_typed(ast, fun) do
+    :erl_syntax_lib.map(
+      fn node ->
+        fun.(:erl_syntax.type(node), node)
+      end,
+      ast
+    )
+  end
+
+  def application_est_to_mfa(application_est) do
+    app_operator =
+      application_est
+      # application_operator fails if we don't revert first
+      |> :erl_syntax.revert()
+      |> :erl_syntax.application_operator()
+
+    case :erl_syntax.type(app_operator) do
+      :module_qualifier ->
+        app_module =
+          app_operator
+          |> :erl_syntax.module_qualifier_argument()
+          |> :erl_syntax.atom_value()
+
+        app_function =
+          app_operator
+          |> :erl_syntax.module_qualifier_body()
+          |> :erl_syntax.atom_value()
+
+        arity =
+          :erl_syntax.application_arguments(application_est)
+          |> Enum.count()
+
+        {:ok, {app_module, app_function, arity}}
+
+      _ ->
+        {:error, {:not_remote, application_est}}
+    end
+  end
 end